updated the information in the README file
