+++ README.md
@@ -35,6 +35,12 @@ python manage.py runserver
 2. Создайте новую учетную запись или войдите существующей.
 3. Создайте новую запись в блоге или оставьте комментарий к существующей.
 
+## Тестирование
+
+Код успешно протестирован с помощью фреймворка Pytest. Запустить тестирование вы сможете через 
+команду pytest, введенную в терминале, либо в PyCharm с помощью команды Run. Процент покрытия 
+тестами кода - 90.
+
 ## Документация:
 
 Дополнительную информацию о структуре проекта и API можно найти в [документации](docs/README.md).
