+++ pyproject.toml
@@ -24,6 +24,7 @@ isort = "^6.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.4.0"
+pytest-cov = "^6.1.1"
 
 [tool.mypy]
 disallow_untyped_defs = true
